Move view components into folder for project clarity
