+++ examples/recipes/src/data_viewer/components/ChartView.tsx
