+++ examples/recipes/src/data_viewer/components/Selector.tsx
